Updated data structure with Single Responsibility in mind
Seperated the SO into two seperate objects, a MeshDataObject and MaterialDataObject, each holds its information in a Value property. This makes the code clearer and follows the Single Responsibility Principal.

Also added a level model I created to show it works with external models that have been imorted into the project.

TODO: Demonstrate multiple materials on a single object.